Some casts

+++ src/post_modules/GTK2/source/support.c
@@ -151,9 +151,9 @@ GdkImage *gdkimage_from_pikeimage(struct object *img, int fast, GdkImage *i) {
       MEMSET(allocated,0,sizeof(allocated));
       for (r=0; r<3; r++) for (g=0; g<4; g++) for (b=0; b<3; b++) {
 	GdkColor color;
-	color.red = (int)(r * (65535/2.0));
-	color.green = (int)(g * (65535/3.0));
-	color.blue = (int)(b * (65535/2.0));
+	color.red = (guint16)(r * (65535/2.0));
+	color.green = (guint16)(g * (65535/3.0));
+	color.blue = (guint16)(b * (65535/2.0));
 	color.pixel = 0;
 	if (gdk_color_alloc(col,&color))
           if (color.pixel<COLORMAP_SIZE)
@@ -161,9 +161,9 @@ GdkImage *gdkimage_from_pikeimage(struct object *img, int fast, GdkImage *i) {
       }
       for (r=0; r<6; r++) for (g=0; g<7; g++) for (b=0; b<6; b++) {
 	GdkColor color;
-	color.red=(int)(r*(65535/5.0));
-	color.green=(int)(g*(65535/6.0));
-	color.blue=(int)(b*(65535/5.0));
+	color.red=(guint16)(r*(65535/5.0));
+	color.green=(guint16)(g*(65535/6.0));
+	color.blue=(guint16)(b*(65535/5.0));
 	color.pixel=0;
 	if (gdk_color_alloc(col,&color))
           if (color.pixel<COLORMAP_SIZE)
